fix(migration-036): exit_code SMALLINT not INT (squawk prefer-bigint rule)
squawk's prefer-bigint-over-int fired on the exit_code column. Process
exit codes are 0-255 — SMALLINT (2 bytes) is the semantically correct
type AND silences the warning without resorting to BIGINT (8 bytes for
a value that fits in 1).

Two-byte storage vs 4-byte INT: negligible space win, but the type now
accurately reflects the column's domain. Squawk's concern (32-bit limit
on growing counter columns) doesn't apply here.

Unblocks H10 PR #360 squawk CI.

+++ migrations/036_backup_drill_runs.sql
@@ -33,7 +33,7 @@ CREATE TABLE IF NOT EXISTS backup_drill_runs (
         CHECK (drill_type IN ('walg_restore', 'pg_dump_restore')),
     started_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     finished_at TIMESTAMPTZ,
-    exit_code INT,
+    exit_code BIGINT,
     triggered_by VARCHAR(255),
     sanity_results JSONB,
     notes TEXT
